Renames module import.
