+++ src/main/java/fr/inria/corese/command/utils/exporter/AbstractExporter.java
@@ -5,8 +5,8 @@ import java.nio.file.Path;
 
 import fr.inria.corese.core.Graph;
 import fr.inria.corese.core.print.ResultFormat;
-import fr.inria.corese.kgram.core.Mappings;
-import fr.inria.corese.sparql.triple.function.extension.ResultFormater;
+import fr.inria.corese.core.kgram.core.Mappings;
+import fr.inria.corese.core.sparql.triple.function.extension.ResultFormater;
 import picocli.CommandLine.Model.CommandSpec;
 
 /**
