+++ src/main/java/fr/inria/corese/command/utils/exporter/rdf/EnumRdfOutputFormat.java
@@ -1,6 +1,6 @@
 package fr.inria.corese.command.utils.exporter.rdf;
 
-import fr.inria.corese.sparql.api.ResultFormatDef;
+import fr.inria.corese.core.sparql.api.ResultFormatDef;
 
 /**
  * Enumeration of output RDF serialization formats.
