+++ src/main/java/fr/inria/corese/command/utils/http/SparqlHttpClient.java
@@ -12,11 +12,11 @@ import fr.inria.corese.command.App;
 import fr.inria.corese.command.utils.TestType;
 import fr.inria.corese.core.Graph;
 import fr.inria.corese.core.query.QueryProcess;
-import fr.inria.corese.kgram.core.Query;
-import fr.inria.corese.sparql.triple.parser.Constant;
-import fr.inria.corese.sparql.triple.update.ASTUpdate;
-import fr.inria.corese.sparql.triple.update.Composite;
-import fr.inria.corese.sparql.triple.update.Update;
+import fr.inria.corese.core.kgram.core.Query;
+import fr.inria.corese.core.sparql.triple.parser.Constant;
+import fr.inria.corese.core.sparql.triple.update.ASTUpdate;
+import fr.inria.corese.core.sparql.triple.update.Composite;
+import fr.inria.corese.core.sparql.triple.update.Update;
 import jakarta.ws.rs.client.Client;
 import jakarta.ws.rs.client.ClientBuilder;
 import jakarta.ws.rs.client.Entity;
